Adds a failing test
The session has MIDI mappings that are used by the pd patch to control the
playback. Pd MIDI output should be routed into Renoise for MIDI Mapping. Renoise
MIDI output should go to Pd.

The test is launched in the pd patch and outputs its results to the pd console.

A Renoise session with 2 patterns. The first one is empty, the second one has a
note on the first line.

We start the playback at pattern 1, wait a moment, and send a QZ:next_pattern
command. Sometimes, the first note of the second pattern is never played, which
is unexpected. It seems to depend on the waiting time between the playback start
and the QZ:next_pattern command : some values always produce a fail, some never,
and some give varying results…